クラスの構成を大幅に見直し

+++ UtilsTest/IdentifiableTest.cpp
@@ -35,14 +35,14 @@ namespace UtilsTest
 		TEST_METHOD(operator_bl_eq3)
 		{
 			Identifiable id1(8);
-			unsigned int id2 = 8;
+			unsigned long id2 = 8;
 			Assert::IsTrue(id1 == id2);
 		}
 
 		TEST_METHOD(operator_bl_eq4)
 		{
 			Identifiable id1(8);
-			unsigned int id2 = 0;
+			unsigned long id2 = 0;
 			Assert::IsFalse(id1 == id2);
 		}
 
@@ -70,7 +70,7 @@ namespace UtilsTest
 		TEST_METHOD(operator_bl_neq4)
 		{
 			Identifiable id1(8);
-			unsigned int id2 = 1;
+			unsigned long id2 = 1;
 			Assert::IsTrue(id1 != id2);
 		}
 
@@ -98,42 +98,42 @@ namespace UtilsTest
 		TEST_METHOD(operator_less4)
 		{
 			Identifiable id1(7);
-			unsigned int id2(8);
+			unsigned long id2(8);
 			Assert::IsTrue(id1 < id2);
 		}
 
 		TEST_METHOD(operator_less5)
 		{
 			Identifiable id1(7);
-			unsigned int id2 = 7;
+			unsigned long id2 = 7;
 			Assert::IsFalse(id1 < id2);
 		}
 
 		TEST_METHOD(operator_less6)
 		{
 			Identifiable id1(7);
-			unsigned int id2 = 6;
+			unsigned long id2 = 6;
 			Assert::IsFalse(id1 < id2);
 		}
 
 		TEST_METHOD(operator_greater1)
 		{
 			Identifiable id1(8);
-			unsigned int id2 = 7;
+			unsigned long id2 = 7;
 			Assert::IsTrue(id1 > id2);
 		}
 
 		TEST_METHOD(operator_greater2)
 		{
 			Identifiable id1(7);
-			unsigned int id2 = 7;
+			unsigned long id2 = 7;
 			Assert::IsFalse(id1 > id2);
 		}
 
 		TEST_METHOD(operator_greater3)
 		{
 			Identifiable id1(6);
-			unsigned int id2 = 7;
+			unsigned long id2 = 7;
 			Assert::IsFalse(id1 > id2);
 		}
 
@@ -182,42 +182,42 @@ namespace UtilsTest
 		TEST_METHOD(operator_leq4)
 		{
 			Identifiable id1(6);
-			unsigned int id2 = 7;
+			unsigned long id2 = 7;
 			Assert::IsTrue(id1 <= id2);
 		}
 
 		TEST_METHOD(operator_leq5)
 		{
 			Identifiable id1(7);
-			unsigned int id2 = 7;
+			unsigned long id2 = 7;
 			Assert::IsTrue(id1 <= id2);
 		}
 
 		TEST_METHOD(operator_leq6)
 		{
 			Identifiable id1(8);
-			unsigned int id2 = 7;
+			unsigned long id2 = 7;
 			Assert::IsFalse(id1 <= id2);
 		}
 
 		TEST_METHOD(operator_geq1)
 		{
 			Identifiable id1(8);
-			unsigned int id2 = 7;
+			unsigned long id2 = 7;
 			Assert::IsTrue(id1 >= id2);
 		}
 
 		TEST_METHOD(operator_geq2)
 		{
 			Identifiable id1(7);
-			unsigned int id2 = 7;
+			unsigned long id2 = 7;
 			Assert::IsTrue(id1 >= id2);
 		}
 
 		TEST_METHOD(operator_geq3)
 		{
 			Identifiable id1(6);
-			unsigned int id2 = 7;
+			unsigned long id2 = 7;
 			Assert::IsFalse(id1 >= id2);
 		}
 
@@ -245,112 +245,112 @@ namespace UtilsTest
 		TEST_METHOD(operator_bl_eq_ptr1)
 		{
 			std::shared_ptr<Identifiable> id1 = std::make_shared<Identifiable>(7);
-			unsigned int id2 = 7;
+			unsigned long id2 = 7;
 			Assert::IsTrue(id1 == id2);
 		}
 
 		TEST_METHOD(operator_bl_eq_ptr2)
 		{
 			std::shared_ptr<Identifiable> id1 = std::make_shared<Identifiable>(7);
-			unsigned int id2 = 8;
+			unsigned long id2 = 8;
 			Assert::IsFalse(id1 == id2);
 		}
 
 		TEST_METHOD(operator_bl_neq_ptr1)
 		{
 			std::shared_ptr<Identifiable> id1 = std::make_shared<Identifiable>(7);
-			unsigned int id2 = 8;
+			unsigned long id2 = 8;
 			Assert::IsTrue(id1 != id2);
 		}
 
 		TEST_METHOD(operator_bl_neq_ptr2)
 		{
 			std::shared_ptr<Identifiable> id1 = std::make_shared<Identifiable>(8);
-			unsigned int id2 = 8;
+			unsigned long id2 = 8;
 			Assert::IsFalse(id1 != id2);
 		}
 
 		TEST_METHOD(operator_less_ptr1) 
 		{
 			std::shared_ptr<Identifiable> id1 = std::make_shared<Identifiable>(7);
-			unsigned int id2 = 8;
+			unsigned long id2 = 8;
 			Assert::IsTrue(id1 < id2);
 		}
 
 		TEST_METHOD(operator_less_ptr2)
 		{
 			std::shared_ptr<Identifiable> id1 = std::make_shared<Identifiable>(8);
-			unsigned int id2 = 8;
+			unsigned long id2 = 8;
 			Assert::IsFalse(id1 < id2);
 		}
 
 		TEST_METHOD(operator_less_ptr3)
 		{
 			std::shared_ptr<Identifiable> id1 = std::make_shared<Identifiable>(9);
-			unsigned int id2 = 8;
+			unsigned long id2 = 8;
 			Assert::IsFalse(id1 < id2);
 		}
 
 		TEST_METHOD(operator_grater_ptr1)
 		{
 			std::shared_ptr<Identifiable> id1 = std::make_shared<Identifiable>(9);
-			unsigned int id2 = 8;
+			unsigned long id2 = 8;
 			Assert::IsTrue(id1 > id2);
 		}
 
 		TEST_METHOD(operator_grater_ptr2)
 		{
 			std::shared_ptr<Identifiable> id1 = std::make_shared<Identifiable>(8);
-			unsigned int id2 = 8;
+			unsigned long id2 = 8;
 			Assert::IsFalse(id1 > id2);
 		}
 
 		TEST_METHOD(operator_grater_ptr3)
 		{
 			std::shared_ptr<Identifiable> id1 = std::make_shared<Identifiable>(7);
-			unsigned int id2 = 8;
+			unsigned long id2 = 8;
 			Assert::IsFalse(id1 > id2);
 		}
 
 		TEST_METHOD(operator_leq_ptr1)
 		{
 			std::shared_ptr<Identifiable> id1 = std::make_shared<Identifiable>(7);
-			unsigned int id2 = 8;
+			unsigned long id2 = 8;
 			Assert::IsTrue(id1 <= id2);
 		}
 
 		TEST_METHOD(operator_leq_ptr2)
 		{
 			std::shared_ptr<Identifiable> id1 = std::make_shared<Identifiable>(8);
-			unsigned int id2 = 8;
+			unsigned long id2 = 8;
 			Assert::IsTrue(id1 <= id2);
 		}
 
 		TEST_METHOD(operator_leq_ptr3)
 		{
 			std::shared_ptr<Identifiable> id1 = std::make_shared<Identifiable>(9);
-			unsigned int id2 = 8;
+			unsigned long id2 = 8;
 			Assert::IsFalse(id1 <= id2);
 		}
 
 		TEST_METHOD(operator_geq_ptr1)
 		{
 			std::shared_ptr<Identifiable> id1 = std::make_shared<Identifiable>(9);
-			unsigned int id2 = 8;
+			unsigned long id2 = 8;
 			Assert::IsTrue(id1 >= id2);
 		}
 
 		TEST_METHOD(operator_geq_ptr2)
 		{
 			std::shared_ptr<Identifiable> id1 = std::make_shared<Identifiable>(8);
-			unsigned int id2 = 8;
+			unsigned long id2 = 8;
 			Assert::IsTrue(id1 >= id2);
 		}
 
 		TEST_METHOD(operator_geq_ptr3)
 		{
 			std::shared_ptr<Identifiable> id1 = std::make_shared<Identifiable>(7);
-			unsigned int id2 = 8;
+			unsigned long id2 = 8;
 			Assert::IsFalse(id1 >= id2);
 		}
